win,pipe: don't silently ignore ERROR_OPERATION_ABORTED
Code to silently ignore this error was added in c42a4ca, under the false
premise that this error is somehow equivalent to EINTR on posix platforms.

This isn't true; ERROR_OPERATION_ABORTED doesn't happen unless the
application explicitly aborts an operation. The only place where we do
expect this to happen (when making a blocking ReadFile call in a thread),
it is dealt with explicitly.

Silently ignoring this error elsewhere could potentially hide bugs, hence
libuv shouldn't do it.